requirements.txt
flask
flask-cors
transformers
torch
Pillow
werkzeug
